// LDC.Atlas.WebApp/Client/app/shared/entities/trade-property-privilege.entity.ts
export class TradePropertyPrivilege {
    buyerCodePrivilege: boolean;
    sellerCodePrivilege: boolean;
    counterPartyPrivilege: boolean;
    commodityPrivilege: boolean;
    cropYearPrivilege: boolean;
    quantityCodePrivilege: boolean;
    quantityPrivilege: boolean;
    quantityContractedPrivilege: boolean;
    contractTermsPrivilege: boolean;
    portTermsPrivilege: boolean;
    arbitrationPrivilege: boolean;
    currencyPrivilege: boolean;
    priceCodePrivilege: boolean;
    contractPricePrivilege: boolean;
    contractValuePrivilege: boolean;
    paymentTermsPrivilege: boolean;
    periodTypePrivilege: boolean;
    fromDatePrivilege: boolean;
    toDatePrivilege: boolean;
    positionTypePrivilege: boolean;
    portOfOriginPrivilege: boolean;
    portOfDestinationPrivilege: boolean;
    businessSectorPrivilege: boolean;
    memoPrivilege: boolean;
    contractIssuedOnPrivilege: boolean;
    otherReferencePrivilege: boolean;
    vesselNamePrivilege: boolean;
    blDatePrivilege: boolean;
    blReferencePrivilege: boolean;
    hasSuperTradePrivilege: boolean;
    invoicingStatusPrivilege: boolean;
}